Evm pallet code clean
